refactor(다리 건너기): 테스트 클래스별 패키지 분류
