+++ src/main/java/bridge/controller/BridgeGameController.java
@@ -4,8 +4,8 @@ import bridge.BridgeRandomNumberGenerator;
 import bridge.domain.BridgeGame;
 import bridge.domain.BridgeMaker;
 import bridge.domain.GameCommand;
-import bridge.domain.GameStatus;
 import bridge.domain.GameResult;
+import bridge.domain.GameStatus;
 import bridge.exception.CannotCrossBridgeException;
 import bridge.view.InputView;
 import bridge.view.OutputView;
