+++ src/test/java/bridge/domain/BridgeGameTest.java
@@ -1,9 +1,8 @@
-package bridge;
+package bridge.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import bridge.domain.BridgeGame;
 import bridge.exception.CannotCrossBridgeException;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
